liferay-portal/master sync

+++ portal-impl/src/com/liferay/portal/resiliency/spi/agent/HttpClientSPIAgent.java
@@ -15,6 +15,7 @@
 package com.liferay.portal.resiliency.spi.agent;
 
 import com.liferay.petra.string.CharPool;
+import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Portlet;
@@ -27,7 +28,6 @@ import com.liferay.portal.kernel.resiliency.spi.agent.SPIAgent;
 import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
 import com.liferay.portal.kernel.servlet.ReadOnlyServletResponse;
 import com.liferay.portal.kernel.util.InetAddressUtil;
-import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.util.PropsValues;
@@ -36,7 +36,6 @@ import java.io.DataInput;
 import java.io.DataInputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
 
 import java.net.InetSocketAddress;
@@ -122,14 +121,15 @@ public class HttpClientSPIAgent implements SPIAgent {
 	}
 
 	@Override
-	public HttpServletRequest prepareRequest(HttpServletRequest request)
+	public HttpServletRequest prepareRequest(
+			HttpServletRequest httpServletRequest)
 		throws IOException {
 
 		SPIAgentRequest spiAgentRequest = SPIAgentRequest.readFrom(
-			request.getInputStream());
+			httpServletRequest.getInputStream());
 
 		HttpServletRequest spiAgentHttpServletRequest =
-			spiAgentRequest.populateRequest(request);
+			spiAgentRequest.populateRequest(httpServletRequest);
 
 		spiAgentHttpServletRequest.setAttribute(
 			WebKeys.SPI_AGENT_REQUEST, spiAgentRequest);
@@ -139,18 +139,20 @@ public class HttpClientSPIAgent implements SPIAgent {
 
 	@Override
 	public HttpServletResponse prepareResponse(
-		HttpServletRequest request, HttpServletResponse response) {
+		HttpServletRequest httpServletRequest,
+		HttpServletResponse httpServletResponse) {
 
 		HttpServletResponse spiAgentHttpServletResponse =
 			new BufferCacheServletResponse(
-				new ReadOnlyServletResponse(response));
+				new ReadOnlyServletResponse(httpServletResponse));
 
-		request.setAttribute(WebKeys.SPI_AGENT_ORIGINAL_RESPONSE, response);
+		httpServletRequest.setAttribute(
+			WebKeys.SPI_AGENT_ORIGINAL_RESPONSE, httpServletResponse);
 
-		Portlet portlet = (Portlet)request.getAttribute(
+		Portlet portlet = (Portlet)httpServletRequest.getAttribute(
 			WebKeys.SPI_AGENT_PORTLET);
 
-		request.setAttribute(
+		httpServletRequest.setAttribute(
 			WebKeys.SPI_AGENT_RESPONSE,
 			new SPIAgentResponse(portlet.getContextName()));
 
@@ -159,7 +161,8 @@ public class HttpClientSPIAgent implements SPIAgent {
 
 	@Override
 	public void service(
-			HttpServletRequest request, HttpServletResponse response)
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
 		throws PortalResiliencyException {
 
 		Socket socket = null;
@@ -167,7 +170,8 @@ public class HttpClientSPIAgent implements SPIAgent {
 		try {
 			socket = borrowSocket();
 
-			SPIAgentRequest spiAgentRequest = new SPIAgentRequest(request);
+			SPIAgentRequest spiAgentRequest = new SPIAgentRequest(
+				httpServletRequest);
 
 			OutputStream outputStream = socket.getOutputStream();
 
@@ -175,16 +179,15 @@ public class HttpClientSPIAgent implements SPIAgent {
 
 			spiAgentRequest.writeTo(registrationReference, outputStream);
 
-			InputStream inputStream = socket.getInputStream();
-
-			DataInputStream dataInputStream = new DataInputStream(inputStream);
+			DataInputStream dataInputStream = new DataInputStream(
+				socket.getInputStream());
 
 			boolean forceCloseSocket = consumeHttpResponseHead(dataInputStream);
 
 			SPIAgentResponse spiAgentResponse = SPIAgentResponse.readFrom(
 				dataInputStream);
 
-			spiAgentResponse.populate(request, response);
+			spiAgentResponse.populate(httpServletRequest, httpServletResponse);
 
 			returnSocket(socket, forceCloseSocket);
 
@@ -209,69 +212,69 @@ public class HttpClientSPIAgent implements SPIAgent {
 
 	@Override
 	public void transferResponse(
-			HttpServletRequest request, HttpServletResponse response,
-			Exception exception)
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse, Exception exception)
 		throws IOException {
 
-		SPIAgentRequest spiAgentRequest = (SPIAgentRequest)request.getAttribute(
-			WebKeys.SPI_AGENT_REQUEST);
+		SPIAgentRequest spiAgentRequest =
+			(SPIAgentRequest)httpServletRequest.getAttribute(
+				WebKeys.SPI_AGENT_REQUEST);
 
-		request.removeAttribute(WebKeys.SPI_AGENT_REQUEST);
+		httpServletRequest.removeAttribute(WebKeys.SPI_AGENT_REQUEST);
 
 		File requestBodyFile = spiAgentRequest.requestBodyFile;
 
-		if (requestBodyFile != null) {
-			if (!requestBodyFile.delete()) {
-				requestBodyFile.deleteOnExit();
-			}
+		if ((requestBodyFile != null) && !requestBodyFile.delete()) {
+			requestBodyFile.deleteOnExit();
 		}
 
 		SPIAgentResponse spiAgentResponse =
-			(SPIAgentResponse)request.getAttribute(WebKeys.SPI_AGENT_RESPONSE);
+			(SPIAgentResponse)httpServletRequest.getAttribute(
+				WebKeys.SPI_AGENT_RESPONSE);
 
-		request.removeAttribute(WebKeys.SPI_AGENT_RESPONSE);
+		httpServletRequest.removeAttribute(WebKeys.SPI_AGENT_RESPONSE);
 
 		if (exception != null) {
 			spiAgentResponse.setException(exception);
 		}
 		else {
 			BufferCacheServletResponse bufferCacheServletResponse =
-				(BufferCacheServletResponse)response;
+				(BufferCacheServletResponse)httpServletResponse;
 
 			spiAgentResponse.captureResponse(
-				request, bufferCacheServletResponse);
+				httpServletRequest, bufferCacheServletResponse);
 		}
 
-		HttpServletResponse originalResponse =
-			(HttpServletResponse)request.getAttribute(
+		HttpServletResponse originalHttpServletResponse =
+			(HttpServletResponse)httpServletRequest.getAttribute(
 				WebKeys.SPI_AGENT_ORIGINAL_RESPONSE);
 
-		request.removeAttribute(WebKeys.SPI_AGENT_ORIGINAL_RESPONSE);
+		httpServletRequest.removeAttribute(WebKeys.SPI_AGENT_ORIGINAL_RESPONSE);
 
-		originalResponse.setContentLength(8);
+		originalHttpServletResponse.setContentLength(8);
 
 		spiAgentResponse.writeTo(
-			registrationReference, originalResponse.getOutputStream());
+			registrationReference,
+			originalHttpServletResponse.getOutputStream());
 	}
 
 	protected Socket borrowSocket() throws IOException {
 		Socket socket = socketBlockingQueue.poll();
 
-		if (socket != null) {
-			if (socket.isClosed() || !socket.isConnected() ||
-				socket.isInputShutdown() || socket.isOutputShutdown()) {
+		if ((socket != null) &&
+			(socket.isClosed() || !socket.isConnected() ||
+			 socket.isInputShutdown() || socket.isOutputShutdown())) {
 
-				try {
-					socket.close();
-				}
-				catch (IOException ioe) {
-					if (_log.isWarnEnabled()) {
-						_log.warn(ioe, ioe);
-					}
-				}
-
-				socket = null;
+			try {
+				socket.close();
 			}
+			catch (IOException ioe) {
+				if (_log.isWarnEnabled()) {
+					_log.warn(ioe, ioe);
+				}
+			}
+
+			socket = null;
 		}
 
 		if (socket == null) {
