+++ portal-impl/src/com/liferay/portlet/documentlibrary/util/XugglerAutoInstallHelper.java
@@ -16,6 +16,7 @@ package com.liferay.portlet.documentlibrary.util;
 
 import com.liferay.petra.process.ProcessException;
 import com.liferay.petra.reflect.ReflectionUtil;
+import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.configuration.Filter;
 import com.liferay.portal.kernel.io.DummyOutputStream;
 import com.liferay.portal.kernel.log.Log;
@@ -24,7 +25,6 @@ import com.liferay.portal.kernel.util.OSDetector;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.SetUtil;
-import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xuggler.Xuggler;
 import com.liferay.portal.kernel.xuggler.XugglerInstallException;
@@ -226,18 +226,18 @@ public class XugglerAutoInstallHelper {
 
 				Matcher matcher = _pattern.matcher(file.getName());
 
-				if (matcher.matches()) {
-					if (jarFiles.contains(matcher.replaceAll("$1$2"))) {
-						file.delete();
+				if (matcher.matches() &&
+					jarFiles.contains(matcher.replaceAll("$1$2"))) {
 
-						iterator.remove();
-					}
+					file.delete();
+
+					iterator.remove();
 				}
 			}
 		}
 
 		URLClassLoader urlClassLoader = new URLClassLoader(
-			urls.toArray(new URL[urls.size()]), null);
+			urls.toArray(new URL[0]), null);
 
 		currentThread.setContextClassLoader(urlClassLoader);
